fix(pkg): honest + ESM-safe browser entry/subpaths, drop stray stream-polyfill export (T5 review)
